Сделал хранение связующего графа

+++ Assets/_GameAssets/Scripts/Navigation/HierarchyGrid/2.0/Structs/Edge.cs
@@ -1,25 +1,31 @@
 
 
+using UnityEngine;
+
 [System.Serializable]
 public struct Edge
 {
     public ushort toVertexId;
-    public int weight;
+    public ushort weight;
     public Path path;
 
     public Edge(ushort toVertexId, int weight)
     {
         this.toVertexId = toVertexId;
-        this.weight = weight;
+        this.weight = (ushort)weight;
         path = new Path(betweenPoints: null);
     }
 
     public Edge(ushort toVertexId, int weight, Path path)
     {
         this.toVertexId = toVertexId;
-        this.weight = weight;
+        this.weight = (ushort)weight;
         this.path = path;
     }
+
+    
+
+
 
 }
